Extract pure setting generation code into new module
