+++ server/server-config-generate/src/main/kotlin/suwayomi/tachidesk/server/settings/generation/GenerateSettingsMain.kt
