+++ settings.gradle.kts
@@ -3,6 +3,7 @@ rootProject.name = System.getenv("ProductName") ?: "Suwayomi-Server"
 include("server")
 include("server:i18n")
 include("server:server-config")
+include("server:server-config-generate")
 
 include("AndroidCompat")
 include("AndroidCompat:Config")
